refactor(command): clean up tpr
